design: 독후감 상세 페이지 디자인 수정 - 독후감 제목 최상단 배치 - 유저 프로필을 독후감 제목 하단으로 배치 - 도서 정보를 독후감 본문과 다른 독후감 중간에 배치 - 좋아요 버튼을 하트 모양으로 변경 - 작성자와 유저가 동일시 점3개(수정,삭제) - 작성자와 유저가 다를시  팔로우/언팔로우 버튼
