+++ src/app/reviews/[id]/page.module.scss
@@ -1,7 +1,6 @@
 // src/app/reviews/[id]/page.module.scss
 .container {
   min-height: calc(100vh - 64px);
-  background: #f9fafb;
   padding: 40px 20px;
 }
