+++ src/components/reviews/detail/review-content.module.scss
@@ -1,176 +1,35 @@
 .container {
   background: white;
-  border-radius: 12px;
-  padding: 32px;
-  margin-bottom: 24px;
-  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.08);
-}
+  padding: 32px 0px 32px 32px;
 
-.header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-bottom: 32px;
-  padding-bottom: 24px;
-  border-bottom: 1px solid #e5e7eb;
-}
-
-.userSection {
-  display: flex;
-  align-items: center;
-  gap: 16px;
-}
-
-.profileImage {
-  width: 48px;
-  height: 48px;
-  border-radius: 50%;
-  object-fit: cover;
-  border: 2px solid #e5e7eb;
-  cursor: pointer;
-  transition: all 0.2s;
-
-  &:hover {
-    border-color: #2d5a3d;
-    transform: scale(1.05);
-  }
-}
-
-.profilePlaceholder {
-  width: 48px;
-  height: 48px;
-  border-radius: 50%;
-  background: linear-gradient(135deg, #007664 0%, #00a389 100%);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 18px;
-  font-weight: 700;
-  color: #ffffff;
-  text-transform: uppercase;
-  cursor: pointer;
-  transition: all 0.2s;
-
-  &:hover {
-    transform: scale(1.05);
-    box-shadow: 0 4px 12px rgba(0, 118, 100, 0.3);
-  }
-}
-
-.userInfo {
-  display: flex;
-  flex-direction: column;
-  gap: 4px;
-}
-
-.userId {
-  font-size: 16px;
-  font-weight: 600;
-  color: #1f2937;
-  margin: 0;
-  cursor: pointer;
-  transition: all 0.2s;
-
-  &:hover {
-    color: #2d5a3d;
-    text-decoration: underline;
-  }
-}
-
-.date {
-  font-size: 14px;
-  color: #9ca3af;
-  margin: 0;
-}
-
-.actions {
-  display: flex;
-  gap: 8px;
-}
-
-.editButton,
-.deleteButton,
-.followButton {
-  padding: 8px 16px;
-  border-radius: 6px;
-  font-size: 14px;
-  font-weight: 500;
-  cursor: pointer;
-  transition: all 0.2s;
-  border: 1px solid #e5e7eb;
-  background: white;
-}
-
-.editButton {
-  color: #2d5a3d;
-  border-color: #2d5a3d;
-
-  &:hover {
-    background: #2d5a3d;
-    color: white;
-  }
-}
-
-.deleteButton {
-  color: #ef4444;
-  border-color: #fee2e2;
-
-  &:hover {
-    background: #fef2f2;
-    border-color: #fecaca;
-  }
-}
-
-.followButton {
-  color: #2d5a3d;
-  border-color: #2d5a3d;
-
-  &:hover:not(:disabled) {
-    background: #2d5a3d;
-    color: white;
-  }
-
-  &.following {
-    border-color: #6b7280;
-    color: #6b7280;
-    background: #f3f4f6;
-
-    &:hover:not(:disabled) {
-      background: #6b7280;
-      color: white;
-    }
-  }
-
-  &:disabled {
-    opacity: 0.5;
-    cursor: not-allowed;
+  @media (max-width: 768px) {
+    padding: 24px 20px;
   }
 }
 
 .content {
-  margin-bottom: 32px;
-}
-
-.title {
-  font-size: 28px;
-  font-weight: 700;
-  color: #1f2937;
-  line-height: 1.4;
-  margin: 0 0 12px 0;
-}
-
-.category {
-  font-size: 14px;
-  color: #2d5a3d;
-  font-weight: 600;
-  margin: 0 0 24px 0;
+  margin: 0;
 }
 
 .body {
-  font-size: 16px;
-  line-height: 1.8;
-  color: #374151;
   word-break: break-word;
+
+  p {
+    line-height: 1.8;
+    font-size: 16px;
+  }
+
+  h1 {
+    font-size: 28px;
+  }
+
+  h2 {
+    font-size: 24px;
+  }
+
+  h3 {
+    font-size: 20px;
+  }
 
   img {
     max-width: 100%;
@@ -217,7 +76,7 @@
     background: #f3f4f6;
     padding: 2px 6px;
     border-radius: 4px;
-    font-size: 14px;
+    font-size: 16px;
     font-family: 'Courier New', monospace;
   }
 
@@ -233,41 +92,6 @@
       background: transparent;
       padding: 0;
       color: inherit;
-    }
-  }
-}
-
-.footer {
-  display: flex;
-  gap: 12px;
-  padding-top: 24px;
-  border-top: 1px solid #e5e7eb;
-}
-
-.likeButton {
-  padding: 10px 20px;
-  border-radius: 8px;
-  font-size: 15px;
-  font-weight: 600;
-  cursor: pointer;
-  transition: all 0.2s;
-  border: 2px solid #2d5a3d;
-  background: white;
-  color: #2d5a3d;
-
-  &:hover {
-    background: #2d5a3d;
-    color: white;
-    transform: translateY(-2px);
-    box-shadow: 0 4px 12px rgba(45, 90, 61, 0.2);
-  }
-
-  &.liked {
-    background: #2d5a3d;
-    color: white;
-
-    &:hover {
-      background: #1f4029;
     }
   }
 }
