+++ src/components/reviews/detail/review-detail.module.scss
@@ -4,89 +4,9 @@
   padding: 0;
   min-height: 100vh;
   position: relative;
-}
-
-.container {
-  max-width: 1200px;
-  margin: 0 auto;
-  padding: 0;
-  background: white;
-  min-height: 100vh;
-  position: relative;
 
   @media (max-width: 768px) {
     max-width: 100%;
-  }
-}
-
-.backButton {
-  position: absolute;
-  top: 120px;
-  right: 40px;
-  display: flex;
-  align-items: center;
-  gap: 6px;
-  padding: 8px 12px;
-  background: rgba(0, 0, 0, 0.05);
-  border: none;
-  border-radius: 20px;
-  font-size: 14px;
-  font-weight: 500;
-  color: #4b5563;
-  cursor: pointer;
-  transition: all 0.2s ease;
-  z-index: 10;
-  backdrop-filter: blur(10px);
-
-  svg {
-    width: 18px;
-    height: 18px;
-    stroke-width: 2.5;
-  }
-
-  &:hover {
-    background: rgba(0, 0, 0, 0.1);
-    color: #1f2937;
-    transform: translateY(-1px);
-    box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-  }
-
-  &:active {
-    transform: translateY(0) scale(0.95);
-  }
-
-  &:focus {
-    outline: none;
-    box-shadow: 0 0 0 3px rgba(59, 130, 246, 0.2);
-  }
-
-  @media (max-width: 1024px) {
-    top: 110px;
-    right: 30px;
-  }
-
-  @media (max-width: 768px) {
-    top: 100px;
-    right: 20px;
-    padding: 6px 10px;
-    font-size: 13px;
-
-    svg {
-      width: 16px;
-      height: 16px;
-    }
-  }
-
-  @media (max-width: 480px) {
-    top: 90px;
-    right: 16px;
-    padding: 5px 8px;
-    font-size: 12px;
-
-    svg {
-      width: 14px;
-      height: 14px;
-    }
   }
 }
 
@@ -94,26 +14,9 @@
   display: flex;
   flex-direction: column;
   gap: 0;
-  padding: 60px 40px 40px;
+  padding: 0px 40px 0px 40px;
   max-width: 1000px;
   margin: 0 auto;
-
-  /* 첫 번째 자식 요소 (BookInfo) */
-  > :first-child {
-    border-radius: 12px 12px 0 0 !important;
-    box-shadow: none !important;
-  }
-
-  /* 마지막 자식 요소 (CommentSection) */
-  > :last-child {
-    border-radius: 0 0 12px 12px !important;
-  }
-
-  /* 중간 요소들 (ReviewContent) */
-  > :not(:first-child):not(:last-child) {
-    border-radius: 0 !important;
-    box-shadow: none !important;
-  }
 
   @media (max-width: 1024px) {
     max-width: 900px;
@@ -123,25 +26,280 @@
   @media (max-width: 768px) {
     max-width: 100%;
     padding: 50px 20px 30px;
-
-    > :first-child {
-      border-radius: 8px 8px 0 0 !important;
-    }
-
-    > :last-child {
-      border-radius: 0 0 8px 8px !important;
-    }
   }
 
   @media (max-width: 480px) {
     padding: 40px 16px 20px;
+  }
+}
 
-    > :first-child {
-      border-radius: 6px 6px 0 0 !important;
+.titleSection {
+  background: white;
+  padding: 32px;
+  border-radius: 12px 12px 0 0;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.08);
+  border-bottom: 1px solid #e5e7eb;
+  width: 100%;
+
+  @media (max-width: 768px) {
+    padding: 24px 20px;
+    border-radius: 8px 8px 0 0;
+  }
+}
+
+.reviewTitle {
+  font-size: 28px;
+  font-weight: 700;
+  color: #1f2937;
+  line-height: 1.4;
+  margin: 0;
+  width: 100%;
+  word-wrap: break-word;
+  word-break: keep-all;
+  overflow-wrap: break-word;
+  text-wrap: wrap;
+
+  @media (max-width: 768px) {
+    font-size: 24px;
+  }
+}
+
+.authorSection {
+  background: white;
+  padding: 24px 32px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  border-bottom: 1px solid #e5e7eb;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.08);
+
+  @media (max-width: 768px) {
+    padding: 20px;
+  }
+}
+
+.authorInfo {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+  flex: 1;
+}
+
+.authorDetails {
+  display: flex;
+  flex-direction: column;
+  gap: 4px;
+}
+
+.authorName {
+  font-size: 16px;
+  font-weight: 600;
+  color: #1f2937;
+  margin: 0;
+  cursor: pointer;
+  transition: all 0.2s;
+
+  &:hover {
+    color: #2d5a3d;
+    text-decoration: underline;
+  }
+}
+
+.date {
+  font-size: 14px;
+  color: #9ca3af;
+  margin: 0;
+}
+
+.menuContainer {
+  position: relative;
+  flex-shrink: 0;
+}
+
+.menuButton {
+  background: none;
+  border: none;
+  cursor: pointer;
+  padding: 4px 8px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  transition: all 0.2s;
+  border-radius: 4px;
+
+  &:hover {
+    background: #f3f4f6;
+  }
+}
+
+.menuIcon {
+  font-size: 20px;
+  color: #6b7280;
+  line-height: 1;
+  display: block;
+  font-weight: bold;
+}
+
+.menuDropdown {
+  position: absolute;
+  top: 100%;
+  right: 0;
+  margin-top: 4px;
+  background: #ffffff;
+  border: 1px solid #e5e7eb;
+  border-radius: 8px;
+  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
+  overflow: hidden;
+  z-index: 10;
+  min-width: 120px;
+}
+
+.menuItem {
+  display: block;
+  width: 100%;
+  padding: 10px 16px;
+  font-size: 14px;
+  color: #374151;
+  background: none;
+  border: none;
+  text-align: left;
+  cursor: pointer;
+  transition: all 0.2s;
+
+  &:hover {
+    background: #f9fafb;
+    color: #007664;
+  }
+
+  &:not(:last-child) {
+    border-bottom: 1px solid #f3f4f6;
+  }
+}
+
+.followButton {
+  padding: 8px 16px;
+  border-radius: 6px;
+  font-size: 14px;
+  font-weight: 500;
+  cursor: pointer;
+  transition: all 0.2s;
+  border: 1px solid #e5e7eb;
+  background: white;
+  color: #2d5a3d;
+  border-color: #2d5a3d;
+
+  &:hover:not(:disabled) {
+    background: #2d5a3d;
+    color: white;
+  }
+
+  &.following {
+    border-color: #6b7280;
+    color: #6b7280;
+    background: #f3f4f6;
+
+    &:hover:not(:disabled) {
+      background: #6b7280;
+      color: white;
+    }
+  }
+
+  &:disabled {
+    opacity: 0.5;
+    cursor: not-allowed;
+  }
+}
+
+.likeSection {
+  background: white;
+  padding: 32px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  border-top: 1px solid #e5e7eb;
+  border-bottom: 1px solid #e5e7eb;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.08);
+
+  @media (max-width: 768px) {
+    padding: 24px 20px;
+  }
+}
+
+.likeButton {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  gap: 12px;
+  background: white;
+  border: 2px solid #e5e7eb;
+  cursor: pointer;
+  padding: 12px 24px;
+  transition: all 0.2s;
+  border-radius: 50px;
+
+  &:hover {
+    border-color: #ef4444;
+    background: #fef2f2;
+    transform: translateY(-2px);
+    box-shadow: 0 4px 12px rgba(239, 68, 68, 0.2);
+  }
+
+  &:active {
+    transform: translateY(0);
+  }
+
+  &.liked {
+    border-color: #ef4444;
+    background: #fef2f2;
+
+    .heartIcon {
+      color: #ef4444;
+      animation: heartBeat 0.3s ease;
     }
 
-    > :last-child {
-      border-radius: 0 0 6px 6px !important;
+    .likeCount {
+      color: #ef4444;
     }
+
+    &:hover {
+      background: #fee2e2;
+    }
+  }
+}
+
+.heartIcon {
+  width: 24px;
+  height: 24px;
+  color: #9ca3af;
+  transition: all 0.2s;
+  stroke-width: 2;
+  flex-shrink: 0;
+
+  @media (max-width: 768px) {
+    width: 22px;
+    height: 22px;
+  }
+}
+
+.likeCount {
+  font-size: 16px;
+  font-weight: 600;
+  color: #6b7280;
+  transition: all 0.2s;
+
+  @media (max-width: 768px) {
+    font-size: 14px;
+  }
+}
+
+@keyframes heartBeat {
+  0% {
+    transform: scale(1);
+  }
+  50% {
+    transform: scale(1.3);
+  }
+  100% {
+    transform: scale(1);
   }
 }
